+++ Tavisca.Bootcamp.LanguageBasics.Exercise1/Enum.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+namespace Tavisca.Bootcamp.LanguageBasics.Exercise1
+{
+    enum EquationParts{
+        Operand1,
+        Operand2,
+        Result,
+        Invalid
+
+    }
+}
